+++ acmed/src/acme_proto/structs.rs
@@ -22,5 +22,5 @@ pub use account::{Account, AccountDeactivation, AccountResponse, AccountUpdate};
 pub use authorization::{Authorization, AuthorizationStatus, Challenge};
 pub use deserialize_from_str;
 pub use directory::Directory;
-pub use error::{AcmeError, HttpApiError};
+pub use error::{AcmeError, ApiError, HttpApiError};
 pub use order::{Identifier, IdentifierType, NewOrder, Order, OrderStatus};