+++ CHANGELOG.md
@@ -13,6 +13,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Unreleased]
+
+### Added
+- ACMEd now displays a warning when the server indicates an error in an order or an authorization.
+
+
 ## [0.4.0] - 2019-05-08
 
 ### Added